uncomment checking for permissions

+++ apps/socials-bot/src/commands/clear-events.ts
@@ -6,13 +6,13 @@ export async function clearEvents(
   db: ReturnType<typeof connect>,
   permissions: Permission[],
 ) {
-  // if (
-  //   !permissions.includes("president") ||
-  //   !permissions.includes("vice-president")
-  // )
-  //   throw new Error(
-  //     "You must be the president or vice president to clear events",
-  //   );
+  if (
+    !permissions.includes("president") ||
+    !permissions.includes("vice-president")
+  )
+    throw new Error(
+      "You must be the president or vice president to clear events",
+    );
 
   return await db.delete(events);
 }
